pyproject.toml
[tool.poetry]
name = "ibis-mssqlops"
version = "0.1.0"
description = ""
authors = ["Marcus Holanda <mvsh777@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.24"
ibis-framework = {extras = ["mssql"], version = "^7.2.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
